rename: 피드백에 따른 Qna,Review 폴더명 qnaAndReview로 수정
