+++ src/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import Logo from 'components/Brand/Logo';
+import Logo from '@/components/Brand/Logo';
 import { Link } from 'react-router-dom';
 
 export default function Footer() {
