+++ src/pages/EditNotice.tsx
@@ -1,7 +1,7 @@
 import PageMainTitle from '@/components/PageMainTitle';
-import FormCkEditor from '@/components/QnA,Review/FormCkEditor';
-import FormTitleInput from '@/components/QnA,Review/FormTitleInput';
-import WriteButton from '@/components/QnA,Review/WriteButton';
+import FormCkEditor from '@/components/qnaAndReview/FormCkEditor';
+import FormTitleInput from '@/components/qnaAndReview/FormTitleInput';
+import WriteButton from '@/components/qnaAndReview/WriteButton';
 import { useData } from '@/store/useData';
 import { useForm } from '@/store/useForm';
 import axiosInstance from '@/utils/axiosInstance';
